refactor: extract provider factories
