+++ packages/router-component-store/src/lib/global-router-store/provide-global-router-store.ts
@@ -1,0 +1,13 @@
+import { ClassProvider, Provider } from '@angular/core';
+import { RouterComponentStore } from '../router-component-store';
+import { GlobalRouterComponentStore } from './global-router-component-store';
+import { GlobalRouterStore } from './global-router-store';
+
+export function provideGlobalRouterStore(): Provider {
+  const globalRouterStoreProvider: ClassProvider = {
+    provide: RouterComponentStore,
+    useClass: GlobalRouterStore,
+  };
+
+  return [globalRouterStoreProvider, GlobalRouterComponentStore];
+}
